rename pytests
