+++ nemo_text_processing/inverse_text_normalization/taggers/cardinal.py
@@ -15,7 +15,13 @@
 
 import pynini
 from nemo_text_processing.inverse_text_normalization.data_loader_utils import get_abs_path
-from nemo_text_processing.inverse_text_normalization.graph_utils import NEMO_DIGIT, NEMO_SIGMA, NEMO_SPACE, GraphFst, delete_space
+from nemo_text_processing.inverse_text_normalization.graph_utils import (
+    NEMO_DIGIT,
+    NEMO_SIGMA,
+    NEMO_SPACE,
+    GraphFst,
+    delete_space,
+)
 from nemo_text_processing.inverse_text_normalization.utils import num_to_word
 from pynini.lib import pynutil
