+++ nemo_text_processing/inverse_text_normalization/taggers/date.py
@@ -14,7 +14,12 @@
 
 import pynini
 from nemo_text_processing.inverse_text_normalization.data_loader_utils import get_abs_path
-from nemo_text_processing.inverse_text_normalization.graph_utils import NEMO_SIGMA, GraphFst, delete_extra_space, delete_space
+from nemo_text_processing.inverse_text_normalization.graph_utils import (
+    NEMO_SIGMA,
+    GraphFst,
+    delete_extra_space,
+    delete_space,
+)
 from pynini.lib import pynutil
