+++ nemo_text_processing/inverse_text_normalization/verbalizers/punctuation.py
@@ -14,7 +14,12 @@
 # limitations under the License.
 
 import pynini
-from nemo_text_processing.inverse_text_normalization.graph_utils import NEMO_CHAR, NEMO_NOT_QUOTE, GraphFst, delete_space
+from nemo_text_processing.inverse_text_normalization.graph_utils import (
+    NEMO_CHAR,
+    NEMO_NOT_QUOTE,
+    GraphFst,
+    delete_space,
+)
 from pynini.lib import pynutil
