+++ nemo_text_processing/inverse_text_normalization/verbalizers/time.py
@@ -14,7 +14,13 @@
 # limitations under the License.
 
 import pynini
-from nemo_text_processing.inverse_text_normalization.graph_utils import NEMO_CHAR, NEMO_DIGIT, GraphFst, delete_space, insert_space
+from nemo_text_processing.inverse_text_normalization.graph_utils import (
+    NEMO_CHAR,
+    NEMO_DIGIT,
+    GraphFst,
+    delete_space,
+    insert_space,
+)
 from pynini.lib import pynutil
