+++ tests/nemo_text_processing/data_inverse_text_normalization/test_cases_cardinal.txt
