+++ tests/nemo_text_processing/data_text_normalization/test_cases_ordinal.txt
