+++ tests/nemo_text_processing/test_boundary.py
@@ -13,15 +13,15 @@
 # limitations under the License.
 
 import pytest
-from nemo_tools.text_denormalization.denormalize import denormalize
+from nemo_text_processing.text_normalization.normalize import normalize
 from parameterized import parameterized
 from utils import parse_test_case_file
 
 
-class TestDecimal:
-    @parameterized.expand(parse_test_case_file('data_text_denormalization/test_cases_decimal.txt'))
+class TestBoundary:
+    @parameterized.expand(parse_test_case_file('data_text_normalization/test_cases_boundary.txt'))
     @pytest.mark.run_only_on('CPU')
     @pytest.mark.unit
-    def test_denorm(self, test_input, expected):
-        pred = denormalize(test_input, verbose=False)
+    def test_norm(self, test_input, expected):
+        pred = normalize(test_input, verbose=False)
         assert pred == expected
