+++ tests/nemo_text_processing/test_date.py
@@ -13,15 +13,15 @@
 # limitations under the License.
 
 import pytest
-from nemo_tools.text_normalization.normalize import normalize
+from nemo_text_processing.inverse_text_normalization.inverse_normalize import inverse_normalize
 from parameterized import parameterized
 from utils import parse_test_case_file
 
 
-class TestBoundary:
-    @parameterized.expand(parse_test_case_file('data_text_normalization/test_cases_boundary.txt'))
+class TestDate:
+    @parameterized.expand(parse_test_case_file('data_text_denormalization/test_cases_date.txt'))
     @pytest.mark.run_only_on('CPU')
     @pytest.mark.unit
-    def test_norm(self, test_input, expected):
-        pred = normalize(test_input, verbose=False)
+    def test_denorm(self, test_input, expected):
+        pred = inverse_normalize(test_input, verbose=False)
         assert pred == expected
