+++ tests/nemo_text_processing/test_ordinal.py
@@ -14,15 +14,23 @@
 
 
 import pytest
-from nemo_tools.text_denormalization.denormalize import denormalize
+from nemo_text_processing.inverse_text_normalization.inverse_normalize import inverse_normalize
+from nemo_text_processing.text_normalization.normalize import normalize
 from parameterized import parameterized
 from utils import parse_test_case_file
 
 
-class TestMoney:
-    @parameterized.expand(parse_test_case_file('data_text_denormalization/test_cases_money.txt'))
+class TestOrdinal:
+    @parameterized.expand(parse_test_case_file('data_text_denormalization/test_cases_ordinal.txt'))
     @pytest.mark.run_only_on('CPU')
     @pytest.mark.unit
     def test_denorm(self, test_input, expected):
-        pred = denormalize(test_input, verbose=False)
+        pred = inverse_normalize(test_input, verbose=False)
+        assert pred == expected
+
+    @parameterized.expand(parse_test_case_file('data_text_normalization/test_cases_ordinal.txt'))
+    @pytest.mark.run_only_on('CPU')
+    @pytest.mark.unit
+    def test_norm(self, test_input, expected):
+        pred = normalize(test_input, verbose=False)
         assert pred == expected
